size_t conversion

+++ test/ref/0212_cast_complex.odin
@@ -3,5 +3,5 @@ package test
 fn0212 :: proc()
 {
 	size : i32; off : i32
-	b : i32 = (cast(size_t) size - cast(size_t) off) * 3
+	b : i32 = (cast(uint) size - cast(uint) off) * 3
 }
